src/transactions/pyproject.toml
[project]
name = "transactions"
version = "0.1.0"
description = "transactions service based on LLM to parse Intent about transaction and send to Metamask application"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.12",
    "logfire>=3.12.0",
    "pydantic>=2.11.1",
    "pydantic-ai>=0.0.52",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
    "web3>=7.10.0",
]
